backend/requirements.txt
fastapi  
uvicorn[standard]  
python-multipart  
pydantic  
langchain  
langchain-openai  
openai  
PyMuPDF
pytesseract  
Pillow
pandas  
numpy  
openpyxl  
python-dateutil  
aiofiles  
celery  
redis  
pytest  
pytest-asyncio  
black  
flake8  